requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-groq==0.1.9
faiss-cpu==1.8.0
sentence-transformers==3.0.1
pymupdf==1.24.10
pypdf==4.3.1
ragas==0.1.21
datasets==2.21.0
mlflow==2.16.0
fastapi==0.114.0
uvicorn==0.30.6
streamlit==1.38.0
python-dotenv==1.0.1
pandas==2.2.2
pydantic==2.8.2
